RagChatBot/requirements.txt
streamlit
openai>=1.0.0
chromadb
python-dotenv
PyPDF2
langchain
langchain-text-splitters